+++ install/upgrade/versions/latest.sh
@@ -6,3 +6,7 @@
 #######                      Place additional commands below.                   #######
 #######################################################################################
 
+if [ -e "/etc/apache2/mods-enabled/status.conf" ]; then
+    echo "(*) Disable Apache2 Server Status Module..."
+    a2dismod status > /dev/null 2>&1
+fi